:art: Release를 Activity와 Order로 분리
